requirements.txt
deeppavlov==1.7.0
torch==2.4.1
spacy
fastapi-cli
uvicorn
torch
transformers
pytorch-crf